% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_newton.R
\name{ls_newton}
\alias{ls_newton}
\title{Least Squares Estimation using Newton's Method}
\usage{
ls_newton(y, X, tolerance = 1e-06, max_iterations = 100)
}
\arguments{
\item{y}{response variable}

\item{X}{design matrix with intercept included}

\item{tolerance}{tolerance level for convergence}

\item{max_iterations}{maximum number of iterations}
}
\value{
an object of class 'lsnewton' with the estimated coefficients
}
\description{
Least Squares Estimation using Newton's Method
}
\examples{
y <- c(1, 2, 3)
X <- matrix(c(1, 1, 1, 1, 2, 3), ncol = 2)
result <- ls_newton(y, X)
print(result)
}
