% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_newton.R
\name{print.lsnewton}
\alias{print.lsnewton}
\title{Print method for lsnewton objects}
\usage{
\method{print}{lsnewton}(x, ...)
}
\arguments{
\item{x}{an object of class 'lsnewton'}

\item{...}{additional arguments affecting the summary produced}
}
\description{
Print method for lsnewton objects
}
