% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsnewton.R
\name{ls_newton}
\alias{ls_newton}
\title{least squares regression based on Newton's algorithm}
\usage{
ls_newton(y, X, tolerance = 1e-06, max_iterations = 100)
}
\arguments{
\item{y}{outcome matrix (n x 1)}

\item{X}{design matrix (n x p)}

\item{tolerance}{tolerance}

\item{max_iterations}{maximum number of iterations allowed}
}
\value{
beta vector of length p
}
\description{
This package implements the Newton's algorithm for solving the least square regression
}
\examples{
# example code
set.seed(123)
n <- 100  # Number of data points
p <- 3    # Number of predictor variables
X <- matrix(rnorm(n * p), ncol = p)  # Create a design matrix with p columns
beta_true <- c(2, 3, 1)  # True coefficients
y <- X \%*\% beta_true + rnorm(n)  # Generate response variable
ls_newton(y, X)
}
